# backend/docs/swagger.yaml
definitions:
  dto.CreateMenuRequest:
    properties:
      name:
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
  dto.UpdateMenuRequest:
    properties:
      name:
        type: string
      order:
        type: integer
      parent_id:
        type: string
    required:
    - name
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /menu:
    get:
      description: Retrieve a list of all menus in a tree structure
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all menus
      tags:
      - Menu
    post:
      consumes:
      - application/json
      description: Create a new menu item
      parameters:
      - description: Menu details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new menu
      tags:
      - Menu
  /menu/{id}:
    delete:
      description: Delete an existing menu item by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete a menu
      tags:
      - Menu
    get:
      description: Retrieve details of a specific menu and its children by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a menu by ID
      tags:
      - Menu
    put:
      consumes:
      - application/json
      description: Update the details (name, order, parent) of an existing menu
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated menu details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update a menu
      tags:
      - Menu
swagger: "2.0"